+++ unikmer/cmd/inter.go
@@ -179,7 +179,7 @@ Tips:
 
 					checkError(writer.Flush())
 					if opt.Verbose {
-						log.Infof("%d k-mers saved", len(m))
+						log.Infof("%d k-mers saved to %s", len(m), outFile)
 					}
 					return flagReturn
 				}
@@ -320,7 +320,7 @@ Tips:
 
 		checkError(writer.Flush())
 		if opt.Verbose {
-			log.Infof("%d k-mers saved", len(m))
+			log.Infof("%d k-mers saved to %s", len(m), outFile)
 		}
 	},
 }