+++ unikmer/cmd/union.go
@@ -238,7 +238,7 @@ Tips:
 
 		checkError(writer.Flush())
 		if opt.Verbose {
-			log.Infof("%d k-mers saved", n)
+			log.Infof("%d k-mers saved to %s", n, outFile)
 		}
 	},
 }